seperated out utils from deployer command logic
